Add files via upload
Electronic Voting Machine (EVM) is a simple electronic device that is used to record the votes in 
place of ballot papers and boxes which were used earlier in conventional voting systems. In the 
conventional voting systems, a voter has to vote his/her favorite candidate by putting the stamp 
against his/her name, folding the ballot paper and putting the paper in the ballot box. This is long, 
time consuming process and very much prone to errors. This situation continued till election scene 
was completely changed by electronic voting machine. In this project entitled “ IoT base Electronic 
Voting Machine” a small, lightweight, low cost EVM machine is designed that can be used in 
election. In this voting machine biometric fingerprint system is implemented because biometric 
identifiers cannot be easily misplaced or shared. The voters have to verify their fingerprint to cast 
their votes. If the fingerprint matched, the system will allow the voter to vote his/her favorite 
candidate. One voter can give only one vote to the candidate. The starting and the ending process 
is encrypted with a password that means only the authority can control the whole process. The 
voting records are stored to a cloud server using IoT so that in case any power failure or technical 
problem the latest result will remain stored in the server. 

Components used-
1. Arduino Uno 
2. ESp8266-01
3. 16x2 LCD display 
4. 4x3 Matrix keypad 
5. JM-101 fingerprint sensor 
6. Buzzer 
7. 18650 Rechargeable battery 
8. 10k potentiometer 
9. 680 ohm resistor